[gaming][preparation][simulation] Fix JIRA # SWBUG-13176 SWS : Le pion "ONG transport" renvoie "terrain difficile" sur Sword et pas sur Scipio
graphics' ElevationAt method now returns a double instead of a short.

Conflicts:
	src/libraries/clients_gui/VisionLine.cpp

+++ src/libraries/clients_gui/GL3DWidget.cpp
@@ -184,7 +184,7 @@ void Gl3dWidget::DrawCross( const Point2f& at, float size /* = -1.f*/, E_Unit un
 // -----------------------------------------------------------------------------
 float Gl3dWidget::ElevationAt( const Point2f& point ) const
 {
-    return ( elevation_.ElevationAt( point ) + 2.f ) * zRatio_;
+    return ( static_cast< float >( elevation_.ElevationAt( point ) ) + 2.f ) * zRatio_;
 }
 
 // -----------------------------------------------------------------------------
